% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprints.R
\name{accDeviation}
\alias{accDeviation}
\title{Deviation of an observed and expected profile pair}
\usage{
accDeviation(plot_data)
}
\arguments{
\item{plot_data}{A \code{data.table} with the columns \code{x},
\code{avg_ins} and \code{type}.}
}
\value{
A numeric value.
}
\description{
Deviation of an observed and expected profile pair
}
\examples{
\dontrun{
accDeviation(plotData)
}
}
