% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviations.R
\name{accDeviationScore}
\alias{accDeviationScore}
\title{Central over flanking insertion density}
\usage{
accDeviationScore(data, breaks = c(-250, -200, -25, 25, 200, 250))
}
\arguments{
\item{data}{A \code{data.table} with the columns \code{x} and
\code{avg_ins}.}

\item{breaks}{Interval borders of the footprint windows.}
}
\value{
A numeric value.
}
\description{
Each interval is divided by its width, so windows of different size
stay comparable.
}
\examples{
prof <- data.table::data.table(x = -250:250, avg_ins = 1)
accDeviationScore(prof)
}
