% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{accDiffPanel}
\alias{accDiffPanel}
\title{Bias corrected footprint panel}
\usage{
accDiffPanel(
  df,
  title,
  colors,
  method = "division",
  flankNorm = 30L,
  plotWindow = 200L
)
}
\arguments{
\item{df}{Footprint data from \code{\link{accFootprintData}}.}

\item{title}{Panel title.}

\item{colors}{Named vector of group colours.}

\item{method}{Either "substraction" or "division".}

\item{flankNorm}{Width of the flanking window used for normalisation.}

\item{plotWindow}{Half width of the plotted window.}
}
\value{
A \code{ggplot} object.
}
\description{
Each group is labelled with its deviation score.
}
\examples{
\dontrun{
accDiffPanel(df, "CTCF", c(a = "red", b = "blue"))
}
}
