% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviations.R
\name{addGCBintoAccessome}
\alias{addGCBintoAccessome}
\title{Insertion density of each GC bin}
\usage{
addGCBintoAccessome(ins, gcdist, ignoreStrand = TRUE)
}
\arguments{
\item{ins}{A \code{GRanges} of insertion sites.}

\item{gcdist}{A \code{GRanges} of the genome wide GC distribution.}

\item{ignoreStrand}{If TRUE, strand information is ignored.}
}
\value{
A \code{matrix} of GC bins and their insertion density.
}
\description{
The total width of the bin is the denominator, so bases without an
insertion count as zero.
}
\examples{
\dontrun{
addGCBintoAccessome(ins, gc_dist)
}
}
