% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviations.R
\name{bpparamFromThreads}
\alias{bpparamFromThreads}
\title{Parallel back-end from a thread count}
\usage{
bpparamFromThreads(threads)
}
\arguments{
\item{threads}{Thread count for parallel processing.}
}
\value{
A \code{BiocParallelParam} object.
}
\description{
Parallel back-end from a thread count
}
\keyword{internal}
