% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromTFR-package.R
\docType{package}
\name{chromTFR-package}
\alias{chromTFR}
\alias{chromTFR-package}
\title{chromTFR: Quantification of Chromatin Accessibility Signatures in TFBS}
\description{
chromTFR quantifies transcription factor footprints in chromatin accessibility data using the annotation of the methylTFR package. Tn5 insertion densities replace methylation levels, and the expected profile of a motif is derived either from the GC bin composition of its binding sites or from a k-mer model of the Tn5 sequence preference. The package computes bias corrected deviation scores per sample and motif, draws the underlying footprints and summarises the deviations as a heatmap.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EpigenomeInformatics/chromTFR}
  \item Report bugs at \url{https://github.com/EpigenomeInformatics/chromTFR/issues}
}

}
\author{
\strong{Maintainer}: Irem B. Gunduz \email{irembgunduz@gmail.com} (\href{https://orcid.org/0000-0003-2641-0916}{ORCID})

Authors:
\itemize{
  \item Fabian Muller \email{fabian.mueller@uni-saarland.de} (\href{https://orcid.org/0000-0001-5809-2321}{ORCID})
}

}
\keyword{internal}
