% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviations.R
\name{computeAccDeviation}
\alias{computeAccDeviation}
\title{Bias corrected deviation of one motif}
\usage{
computeAccDeviation(
  motif,
  ins,
  tf_bindsites,
  gcfreqs,
  binIns,
  enhancer = NULL,
  ignoreStrand = TRUE
)
}
\arguments{
\item{motif}{Motif name.}

\item{ins}{A \code{GRanges} of insertion sites.}

\item{tf_bindsites}{A \code{GRangesList} of binding site positions.}

\item{gcfreqs}{A \code{list} of GC bin frequency tables.}

\item{binIns}{GC bin densities of the sample.}

\item{enhancer}{Optional \code{GRanges} restricting the sites.}

\item{ignoreStrand}{If TRUE, strand information is ignored.}
}
\value{
A \code{data.table} with the deviation and the expected
deviation.
}
\description{
Bias corrected deviation of one motif
}
\examples{
\dontrun{
computeAccDeviation("CTCF", ins, tf_bindsites, gcfreqs, binIns)
}
}
