% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviations.R
\name{computeAccExpectations}
\alias{computeAccExpectations}
\title{Expected insertion profile from the GC composition}
\usage{
computeAccExpectations(binIns, gcfreq)
}
\arguments{
\item{binIns}{GC bin densities from \code{\link{addGCBintoAccessome}}.}

\item{gcfreq}{A \code{matrix} of GC bin frequencies of one motif.}
}
\value{
A \code{data.table} with the columns \code{x} and
\code{avg_ins}.
}
\description{
Expected insertion profile from the GC composition
}
\examples{
\dontrun{
computeAccExpectations(binIns, gcfreqs[["CTCF"]])
}
}
