% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviations.R
\name{computeRowZScore}
\alias{computeRowZScore}
\title{Row-wise z-score of a matrix}
\usage{
computeRowZScore(mat)
}
\arguments{
\item{mat}{A \code{matrix}.}
}
\value{
A \code{matrix} of row-wise z-scores.
}
\description{
Row-wise z-score of a matrix
}
\examples{
computeRowZScore(matrix(c(1, 2, 3, 4), nrow = 2))
}
