% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsatac.R
\name{.requireChrAccR}
\alias{.requireChrAccR}
\title{Check that ChrAccR is installed}
\usage{
.requireChrAccR()
}
\value{
Invisible TRUE, called for the error it raises.
}
\description{
Check that ChrAccR is installed
}
\keyword{internal}
