% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsatac.R
\name{getAccRegions}
\alias{getAccRegions}
\title{Region set of a ChrAccR dataset}
\usage{
getAccRegions(dsa, regionType = "peaks", extend = 500)
}
\arguments{
\item{dsa}{A \code{DsATAC} object.}

\item{regionType}{Region type stored in the dataset.}

\item{extend}{Number of bases added on each side.}
}
\value{
A \code{GRanges} object.
}
\description{
Regions are extended so that they still cover the motif flanks.
}
\examples{
\dontrun{
getAccRegions(dsa, regionType = ".peaks.cons", extend = 500)
}
}
