% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsatac.R
\name{getAccSampleAnnotation}
\alias{getAccSampleAnnotation}
\title{Sample annotation of a ChrAccR dataset}
\usage{
getAccSampleAnnotation(dsa)
}
\arguments{
\item{dsa}{A \code{DsATAC} object.}
}
\value{
A \code{data.frame} with one row per sample.
}
\description{
Sample annotation of a ChrAccR dataset
}
\examples{
\dontrun{
getAccSampleAnnotation(dsa)
}
}
