% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsatac.R
\name{getAccSamples}
\alias{getAccSamples}
\title{Sample identifiers of a ChrAccR dataset}
\usage{
getAccSamples(dsa)
}
\arguments{
\item{dsa}{A \code{DsATAC} object.}
}
\value{
A character vector of sample identifiers.
}
\description{
Sample identifiers of a ChrAccR dataset
}
\examples{
\dontrun{
getAccSamples(dsa)
}
}
