% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer.R
\name{kmerBiasProfile}
\alias{kmerBiasProfile}
\title{Expected cut propensity along a set of binding sites}
\usage{
kmerBiasProfile(tfbs, genome, bias, k = 6L, max.sites = 25000)
}
\arguments{
\item{tfbs}{A \code{GRanges} of binding sites, prepared with
\code{prepareTFBS}.}

\item{genome}{A \code{BSgenome} object.}

\item{bias}{Named k-mer weights from \code{\link{computeKmerBias}}.}

\item{k}{Length of the k-mer.}

\item{max.sites}{Number of binding sites drawn for the estimate.}
}
\value{
A \code{data.table} with the columns \code{x} and \code{w}.
}
\description{
Expected cut propensity along a set of binding sites
}
\examples{
\dontrun{
kmerBiasProfile(tfbs, Hsapiens, bias)
}
}
