% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsatac.R
\name{loadAccDataset}
\alias{loadAccDataset}
\title{Load a ChrAccR dataset}
\usage{
loadAccDataset(path)
}
\arguments{
\item{path}{Path of the ChrAccR dataset directory.}
}
\value{
A \code{DsATAC} object.
}
\description{
Load a ChrAccR dataset
}
\examples{
\dontrun{
dsa <- loadAccDataset("dsATAC_filtered")
}
}
