% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprints.R
\name{normaliseAccFlank}
\alias{normaliseAccFlank}
\title{Set the outer flanks of a corrected profile to the baseline}
\usage{
normaliseAccFlank(data, method, flankNorm)
}
\arguments{
\item{data}{The corrected profile as a \code{data.table}.}

\item{method}{Either "substraction" or "division".}

\item{flankNorm}{Width of the flanking window used for normalisation.}
}
\value{
A \code{data.table}.
}
\description{
Set the outer flanks of a corrected profile to the baseline
}
\keyword{internal}
