% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprints.R
\name{plotAccMotifFootprint}
\alias{plotAccMotifFootprint}
\title{Bias corrected footprint of one motif}
\usage{
plotAccMotifFootprint(
  motif,
  tf_bindsites,
  ins,
  sample_name = NULL,
  gc_dist,
  gcfreqs,
  enhancer = NULL,
  ignoreStrand = TRUE,
  method = "division",
  flankNorm = 50,
  smooth = 5L,
  plotWindow = 200L
)
}
\arguments{
\item{motif}{Motif name.}

\item{tf_bindsites}{A \code{GRangesList} of binding site positions.}

\item{ins}{A \code{GRanges} of insertion sites.}

\item{sample_name}{Optional sample label.}

\item{gc_dist}{A \code{GRanges} of the genome wide GC distribution.}

\item{gcfreqs}{A \code{list} of GC bin frequency tables.}

\item{enhancer}{Optional \code{GRanges} restricting the sites.}

\item{ignoreStrand}{If TRUE, strand information is ignored.}

\item{method}{Either "substraction" or "division".}

\item{flankNorm}{Width of the flanking window used for normalisation.}

\item{smooth}{Width of the running mean applied to the profiles.}

\item{plotWindow}{Half width of the plotted window.}
}
\value{
A \code{ggplot} object.
}
\description{
Bias corrected footprint of one motif
}
\examples{
\dontrun{
plotAccMotifFootprint("CTCF", tf_bindsites, ins,
    gc_dist = gc_dist, gcfreqs = gcfreqs)
}
}
