% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprints.R
\name{plotExpectedAccFootprint}
\alias{plotExpectedAccFootprint}
\title{Observed and expected profile of one motif in one sample}
\usage{
plotExpectedAccFootprint(
  motif,
  tf_bindsites,
  ins,
  sample_name = NULL,
  gc_dist,
  gcfreqs,
  enhancer = NULL,
  ignoreStrand = TRUE,
  smooth = 5L,
  returnPlotData = FALSE
)
}
\arguments{
\item{motif}{Motif name.}

\item{tf_bindsites}{A \code{GRangesList} of binding site positions.}

\item{ins}{A \code{GRanges} of insertion sites.}

\item{sample_name}{Optional sample label.}

\item{gc_dist}{A \code{GRanges} of the genome wide GC distribution.}

\item{gcfreqs}{A \code{list} of GC bin frequency tables.}

\item{enhancer}{Optional \code{GRanges} restricting the sites.}

\item{ignoreStrand}{If TRUE, strand information is ignored.}

\item{smooth}{Width of the running mean applied to the profiles.}

\item{returnPlotData}{If TRUE, the plot data is returned as well.}
}
\value{
A \code{ggplot} object, or a list with the plot and its data.
}
\description{
Observed and expected profile of one motif in one sample
}
\examples{
\dontrun{
plotExpectedAccFootprint("CTCF", tf_bindsites, ins,
    gc_dist = gc_dist, gcfreqs = gcfreqs)
}
}
