% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviations.R
\name{prepareTFBS}
\alias{prepareTFBS}
\title{Prepare the binding sites of one motif}
\usage{
prepareTFBS(tfbs, enhancer = NULL, ignoreStrand = TRUE)
}
\arguments{
\item{tfbs}{A \code{GRanges} of binding sites of one motif.}

\item{enhancer}{Optional \code{GRanges} restricting the sites.}

\item{ignoreStrand}{If TRUE, strand information is ignored.}
}
\value{
A \code{GRanges} object.
}
\description{
Prepare the binding sites of one motif
}
\examples{
\dontrun{
prepareTFBS(tf_bindsites[["CTCF"]])
}
}
