% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer.R
\name{profileFromBias}
\alias{profileFromBias}
\title{Expected profile of one sample from stored k-mer identities}
\usage{
profileFromBias(prof, bias)
}
\arguments{
\item{prof}{Output of \code{\link{kmerIndexProfile}}.}

\item{bias}{Named k-mer weights of the sample.}
}
\value{
A \code{data.table} with the columns \code{x} and
\code{avg_ins}.
}
\description{
Expected profile of one sample from stored k-mer identities
}
\examples{
\dontrun{
profileFromBias(prof, bias[["sample_1"]])
}
}
