% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprints.R
\name{smoothProfile}
\alias{smoothProfile}
\title{Running mean of an insertion profile}
\usage{
smoothProfile(data, window = 5L)
}
\arguments{
\item{data}{A \code{data.table} with the columns \code{x} and
\code{avg_ins}.}

\item{window}{Width of the running mean, 1 to switch it off.}
}
\value{
A \code{data.table}.
}
\description{
Running mean of an insertion profile
}
\keyword{internal}
