% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviations.R
\name{validAccMotifs}
\alias{validAccMotifs}
\title{Drop motifs without binding sites or frequency table}
\usage{
validAccMotifs(tf_bindsites, gcfreqs)
}
\arguments{
\item{tf_bindsites}{A \code{GRangesList} of binding site positions.}

\item{gcfreqs}{A \code{list} of GC bin frequency tables.}
}
\value{
A character vector of motif names.
}
\description{
Drop motifs without binding sites or frequency table
}
\keyword{internal}
